+++ src/domain/transaction/stellar/prepare/operations.rs
@@ -83,3 +83,292 @@ where
     // The signer will build the envelope from operations and sign it
     sign_stellar_transaction(signer, stellar_data_with_sim).await
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::{
+        domain::{SignTransactionResponse, SignTransactionResponseStellar},
+        models::{
+            AssetSpec, DecoratedSignature, NetworkTransactionData, NetworkType, OperationSpec,
+            TransactionInput, TransactionStatus,
+        },
+        repositories::MockTransactionCounterTrait,
+        services::{MockSigner, MockStellarProviderTrait},
+    };
+    use soroban_rs::stellar_rpc_client::SimulateTransactionResponse;
+    use soroban_rs::xdr::{self};
+
+    fn create_test_transaction() -> TransactionRepoModel {
+        TransactionRepoModel {
+            id: "test-tx-1".to_string(),
+            relayer_id: "test-relayer".to_string(),
+            status: TransactionStatus::Pending,
+            status_reason: None,
+            network_data: NetworkTransactionData::Stellar(create_test_stellar_data()),
+            created_at: chrono::Utc::now().to_rfc3339(),
+            sent_at: None,
+            confirmed_at: None,
+            valid_until: None,
+            priced_at: None,
+            hashes: vec![],
+            network_type: NetworkType::Stellar,
+            noop_count: None,
+            is_canceled: Some(false),
+        }
+    }
+
+    fn create_test_stellar_data() -> StellarTransactionData {
+        StellarTransactionData {
+            source_account: "GAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAWHF".to_string(),
+            network_passphrase: "Test SDF Network ; September 2015".to_string(),
+            fee: None,
+            sequence_number: None,
+            transaction_input: TransactionInput::Operations(vec![OperationSpec::Payment {
+                destination: "GAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAWHF".to_string(),
+                amount: 10000000, // 1 XLM in stroops
+                asset: AssetSpec::Native,
+            }]),
+            memo: None,
+            valid_until: None,
+            signatures: vec![],
+            hash: None,
+            simulation_transaction_data: None,
+            signed_envelope_xdr: None,
+        }
+    }
+
+    fn create_dummy_signature() -> DecoratedSignature {
+        DecoratedSignature {
+            hint: xdr::SignatureHint([0, 1, 2, 3]),
+            signature: xdr::Signature(vec![0; 64].try_into().unwrap()),
+        }
+    }
+
+    #[tokio::test]
+    async fn test_process_operations_payment_success() {
+        let relayer_id = "test-relayer";
+        let relayer_address = "GAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAWHF";
+
+        let mut counter = MockTransactionCounterTrait::new();
+        counter.expect_get_and_increment().returning(|_, _| Ok(42));
+
+        let provider = MockStellarProviderTrait::new();
+
+        let mut signer = MockSigner::new();
+        signer.expect_sign_transaction().returning(|_| {
+            Box::pin(async {
+                Ok(SignTransactionResponse::Stellar(
+                    SignTransactionResponseStellar {
+                        signature: create_dummy_signature(),
+                    },
+                ))
+            })
+        });
+
+        let tx = create_test_transaction();
+        let stellar_data = create_test_stellar_data();
+
+        let result = process_operations(
+            &counter,
+            relayer_id,
+            relayer_address,
+            &tx,
+            stellar_data,
+            &provider,
+            &signer,
+        )
+        .await;
+
+        match result {
+            Ok(updated_data) => {
+                assert_eq!(updated_data.sequence_number, Some(42));
+                assert_eq!(updated_data.signatures.len(), 1);
+            }
+            Err(e) => panic!("Test failed with error: {:?}", e),
+        }
+    }
+
+    #[tokio::test]
+    async fn test_process_operations_with_soroban_simulation() {
+        let relayer_id = "test-relayer";
+        let relayer_address = "GAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAWHF";
+
+        let mut counter = MockTransactionCounterTrait::new();
+        counter.expect_get_and_increment().returning(|_, _| Ok(42));
+
+        let mut provider = MockStellarProviderTrait::new();
+        // Mock simulation response for Soroban operations
+        provider
+            .expect_simulate_transaction_envelope()
+            .returning(|_| {
+                Box::pin(async {
+                    Ok(SimulateTransactionResponse {
+                        min_resource_fee: 100,
+                        transaction_data: "AAAAAQAAAAAAAAACAAAAAAAAAAAAAAAAAAAABgAAAAEAAAAGAAAAAG0JZTO9fU6p3NeJp5w3TpKhZmx6p1pR7mq9wFwCnEIuAAAAFAAAAAEAAAAAAAAAB8NVb2IAAAH0AAAAAQAAAAAAABfAAAAAAAAAAPUAAAAAAAAENgAAAAA=".to_string(),
+                        ..Default::default()
+                    })
+                })
+            });
+
+        let mut signer = MockSigner::new();
+        signer.expect_sign_transaction().returning(|_| {
+            Box::pin(async {
+                Ok(SignTransactionResponse::Stellar(
+                    SignTransactionResponseStellar {
+                        signature: create_dummy_signature(),
+                    },
+                ))
+            })
+        });
+
+        let tx = create_test_transaction();
+        let mut stellar_data = create_test_stellar_data();
+        // Change to a Soroban operation to trigger simulation
+        stellar_data.transaction_input =
+            TransactionInput::Operations(vec![OperationSpec::InvokeContract {
+                contract_address: "CA7QYNF7SOWQ3GLR2BGMZEHXAVIRZA4KVWLTJJFC7MGXUA74P7UJUWDA"
+                    .to_string(),
+                function_name: "transfer".to_string(),
+                args: vec![],
+                auth: None,
+            }]);
+
+        let result = process_operations(
+            &counter,
+            relayer_id,
+            relayer_address,
+            &tx,
+            stellar_data,
+            &provider,
+            &signer,
+        )
+        .await;
+
+        match result {
+            Ok(updated_data) => {
+                assert_eq!(updated_data.sequence_number, Some(42));
+                assert!(updated_data.simulation_transaction_data.is_some());
+                assert_eq!(updated_data.signatures.len(), 1);
+            }
+            Err(e) => panic!("Test failed with error: {:?}", e),
+        }
+    }
+
+    #[tokio::test]
+    async fn test_process_operations_sequence_failure() {
+        let relayer_id = "test-relayer";
+        let relayer_address = "GAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAWHF";
+
+        let mut counter = MockTransactionCounterTrait::new();
+        counter.expect_get_and_increment().returning(|_, _| {
+            Err(crate::repositories::TransactionCounterError::NotFound(
+                "Counter not found".to_string(),
+            ))
+        });
+
+        let provider = MockStellarProviderTrait::new();
+        let signer = MockSigner::new();
+
+        let tx = create_test_transaction();
+        let stellar_data = create_test_stellar_data();
+
+        let result = process_operations(
+            &counter,
+            relayer_id,
+            relayer_address,
+            &tx,
+            stellar_data,
+            &provider,
+            &signer,
+        )
+        .await;
+
+        assert!(result.is_err());
+        // The sequence error is mapped to UnexpectedError in get_next_sequence
+        assert!(matches!(
+            result.unwrap_err(),
+            TransactionError::UnexpectedError(_)
+        ));
+    }
+
+    #[tokio::test]
+    async fn test_process_operations_build_envelope_failure() {
+        let relayer_id = "test-relayer";
+        let relayer_address = "GAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAWHF";
+
+        let mut counter = MockTransactionCounterTrait::new();
+        counter.expect_get_and_increment().returning(|_, _| Ok(42));
+
+        let provider = MockStellarProviderTrait::new();
+        let mut signer = MockSigner::new();
+        // Empty operations might still reach the signer
+        signer.expect_sign_transaction().returning(|_| {
+            Box::pin(async {
+                Err(crate::models::SignerError::SigningError(
+                    "Cannot sign empty transaction".to_string(),
+                ))
+            })
+        });
+
+        let tx = create_test_transaction();
+        let mut stellar_data = create_test_stellar_data();
+        // Set invalid operation to cause envelope build failure
+        stellar_data.transaction_input = TransactionInput::Operations(vec![]);
+
+        let result = process_operations(
+            &counter,
+            relayer_id,
+            relayer_address,
+            &tx,
+            stellar_data,
+            &provider,
+            &signer,
+        )
+        .await;
+
+        assert!(result.is_err());
+        // Empty operations should cause an error
+        // The exact error depends on where validation happens
+    }
+
+    #[tokio::test]
+    async fn test_process_operations_signer_failure() {
+        let relayer_id = "test-relayer";
+        let relayer_address = "GAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAWHF";
+
+        let mut counter = MockTransactionCounterTrait::new();
+        counter.expect_get_and_increment().returning(|_, _| Ok(42));
+
+        let provider = MockStellarProviderTrait::new();
+
+        let mut signer = MockSigner::new();
+        signer.expect_sign_transaction().returning(|_| {
+            Box::pin(async {
+                Err(crate::models::SignerError::SigningError(
+                    "Signing failed".to_string(),
+                ))
+            })
+        });
+
+        let tx = create_test_transaction();
+        let stellar_data = create_test_stellar_data();
+
+        let result = process_operations(
+            &counter,
+            relayer_id,
+            relayer_address,
+            &tx,
+            stellar_data,
+            &provider,
+            &signer,
+        )
+        .await;
+
+        assert!(result.is_err());
+        assert!(matches!(
+            result.unwrap_err(),
+            TransactionError::SignerError(_)
+        ));
+    }
+}
